Create CCValid
This program uses the Luhn algorithm to validate a credit card number. The Luhn algorithm works by summing up the digits of the card number and checking if the sum is divisible by 10. If it is, then the card number is valid.

To use this program, you can enter a credit card number when prompted. The program will then output whether the credit card number is valid or not.